% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{handle_mergeplot}
\alias{handle_mergeplot}
\title{Merge samplebase information and stationplot data.}
\usage{
handle_mergeplot(samplebase, stationplot)
}
\arguments{
\item{samplebase}{One element of list sample_base, like "sample_base$citygroup2".}

\item{stationplot}{The dataframe station_plot.}
}
\value{
A dataframe contains station_plot information and attribute information, which can be directly used
by geo_pointplot().
}
\description{
This function can merge one attribute from "samplebase" list with "station_plot" dataframe.
}
\details{
This function aims to quickly plot points samplebase. It can be used before geo_pointplot function.
You can x <- handle_mergeplot(samplebase$citygroup2,station_plot), and then geo_pointplot(x,na.rm=T,type=T).
The attribute information contains in "type" column.
}
