% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{handle_splitatts}
\alias{handle_splitatts}
\title{Split attributes from station_atts}
\usage{
handle_splitatts(stationatts, attname, startpoint)
}
\arguments{
\item{stationatts}{The list: station_atts.}

\item{attname}{The attribute you want to split, like character "Cheng Shi Qun".}

\item{startpoint}{The startpoint of the first str behind "(".}
}
\value{
A dataframe contains 2 columns, "V1" save attributes content, "index" save station_index.
}
\description{
This function helps split attributes from station_atts.Rdata
}
\details{
File "station_atts.Rdata" is handled by strsplit function from the big-table with all station-attributes.
"station_atts" is a big list. Each element is a station. This function receives attname and startpoint of "("
to extract the content of attributes.
}
