% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_support.R
\name{merge_outcome}
\alias{merge_outcome}
\title{Merge 3 columns of outcome}
\usage{
merge_outcome(now, previous = NULL, last = NULL, bywhat)
}
\arguments{
\item{now}{The report month, like 2018/06.}

\item{previous}{One year ago of report month, like 2017/06. Or caculated by function caculate_increaseratio.}

\item{last}{One month ago of report month, like 2018/05. Or caculated by function caculate_increaseratio.}

\item{bywhat}{The theme name, like "citygroup" or "portroad".}
}
\description{
This function can merge 3 column-like dataframes into 1 dataframe with 3 columns.
}
\details{
During report writing, merging now-value, previous-value, last-value into 1 dataframe is quite
useful and concise. This function receive 3 column-like dataframes. Each column-like dataframes
has two columns, one is "theme" like "citygroup" or "portroad". These 3 dataframes has the same theme.
The other is Wmean. It's probably caculated by function caculate_carsmean or caculate_increaseratio.
}
