% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_support.R
\name{result_present3}
\alias{result_present3}
\title{Caculate now_wmean previous_increaseratio and last_increaseratio}
\usage{
result_present3(newsjd, previoussjd, lastsjd, attsname, carstype)
}
\arguments{
\item{newsjd}{The jd dataframe which used to be jdnews.}

\item{previoussjd}{The jd dataframe which used to be jdpreviouss.}

\item{lastsjd}{The jd dataframe which used to be jdlasts.}

\item{attsname}{A string that you want to analysis, for example "citygroup2".}

\item{carstype}{A string must be one from "cars", "passcars", "frecars".}
}
\value{
A dataframe consists of 3 columns, including now_wmean, previous_increaseratio, last_increaseratio.
}
\description{
This function can caculate 3 components in one time by given attsname and carstype.
}
\details{
This function serves for report wirting. It's very common that combines now_wmean,
previous_increaseratio and last_increaseratio in one dataframe. This function
is a combination machine which combines most important functions into one.

It's the most useful function in this package.
}
